feature/admin admin_score to use float64

+++ pkg/projects/project_request.go
@@ -318,7 +318,7 @@ type ProjectReviewer struct {
 type AdminUpdateProjectRequest struct {
 	ProjectStatusPrimary   string     `json:"projectStatusPrimary,omitempty"`
 	ProjectStatusSecondary string     `json:"projectStatusSecondary,omitempty"`
-	AdminScore             *int       `json:"adminScore,omitempty"`
+	AdminScore             *float64   `json:"adminScore,omitempty"`
 	FundApprovedAmount     *int64     `json:"fundApprovedAmount,omitempty"`
 	AdminComment           *string    `json:"adminComment,omitempty"`
 	AdminApprovedAt        *time.Time `json:"adminApprovedAt,omitempty"`
